Limit mesh material pipeline to opaque blending
This is a decision following the major engines e.g unity,
unreal e.t.c having only fixed number of blend modes on
meshed material passes.This will limit it to only opaque
meshes but will be extended for other supported types.
Also placing the combinitorial states on the material
pipeline key is not ideal.